requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
python-dotenv
together